moved a selector to content from highlights
