+++ src/app/content/highlights/components/SummaryPopup/ChapterFilter.tsx
@@ -6,9 +6,10 @@ import Checkbox from '../../../../components/Checkbox';
 import { textStyle } from '../../../../components/Typography/base';
 import { match, not } from '../../../../fpUtils';
 import theme from '../../../../theme';
+import { highlightLocationFilters } from '../../../selectors';
 import { filters } from '../../../styles/PopupConstants';
 import { setSummaryFilters } from '../../actions';
-import { highlightLocationFilters, highlightLocationFiltersWithContent, summaryLocationFilters } from '../../selectors';
+import { highlightLocationFiltersWithContent, summaryLocationFilters } from '../../selectors';
 import ColorIndicator from '../ColorIndicator';
 
 interface Props {
