+++ src/app/content/highlights/components/SummaryPopup/FiltersList.tsx
@@ -9,9 +9,10 @@ import { textStyle } from '../../../../components/Typography';
 import { match, not } from '../../../../fpUtils';
 import theme from '../../../../theme';
 import { disablePrint } from '../../../components/utils/disablePrint';
+import { highlightLocationFilters } from '../../../selectors';
 import { splitTitleParts } from '../../../utils/archiveTreeUtils';
 import { setSummaryFilters } from '../../actions';
-import { highlightLocationFilters, summaryColorFilters, summaryLocationFilters } from '../../selectors';
+import { summaryColorFilters, summaryLocationFilters } from '../../selectors';
 
 // tslint:disable-next-line: variable-name
 export const StyledPlainButton = styled(PlainButton)`
