+++ src/app/content/highlights/hooks/initializeMyHighlightsSummary.ts
@@ -3,9 +3,9 @@ import { ActionHookBody } from '../../../types';
 import { actionHook, assertDefined } from '../../../utils';
 import { summaryPageSize } from '../../constants';
 import * as selectContent from '../../selectors';
+import { highlightLocationFilters } from '../../selectors'
 import { extractTotalCounts } from '../../utils/sharedHighlightsUtils';
 import { initializeMyHighlightsSummary, receiveHighlightsTotalCounts, receiveSummaryHighlights } from '../actions';
-import * as select from '../selectors';
 import { loadMoreMyHighlights } from './loadMore';
 
 export const hookBody: ActionHookBody<typeof initializeMyHighlightsSummary> = (services) => async() => {
@@ -14,7 +14,7 @@ export const hookBody: ActionHookBody<typeof initializeMyHighlightsSummary> = (s
 
   // this can only be undefined in dev if you press the MH button before book is loaded, whatever
   const book = assertDefined(selectContent.book(state), 'book should be defined');
-  const locationFilters = select.highlightLocationFilters(state);
+  const locationFilters = highlightLocationFilters(state);
 
   const totalCounts = await highlightClient.getHighlightsSummary({
     scopeId: book.id,
